+++ open/open.go
@@ -9,10 +9,10 @@ package main
 // 启动类
 func main() {
 	//hs.Mysql2Redis()
-	//getHs(1, 11, 1, platfrom_paoyou)
-	//getHs(1, 31, 2, platfrom_li5apuu7)
-	//getHs(1, 61, 1, platfrom_madou)
-	//getHs(1, 11, 1, platfrom_maomi)
+	//getHs(1, 4, 1, platfrom_paoyou)
+	//getHs(1, 11, 1, platfrom_li5apuu7)
+	getHs(1, 100, 1, platfrom_madou)
+	//getHs(1, 9, 1, platfrom_maomi)
 	// G. max page 25865 / 20 + 1 ==1294
-	getHs(1, 1294, 2, platfrom_G)
+	//getHs(1, 601, 20, platfrom_G)
 }
