+++ src/f104.go
@@ -1,11 +1,6 @@
 package src
 
 import (
-	"fmt"
-	"github.com/gin-gonic/gin"
-	"httpParse/hs"
-	"net/http"
-	"strconv"
 	"time"
 )
 
@@ -21,19 +16,4 @@ func TimeTask() {
 		ch <- 1
 	}()
 	<-ch
-}
-
-func SaveInfo(c *gin.Context) {
-	page, _ := strconv.Atoi(c.Param("page"))
-	start, _ := strconv.Atoi(c.Param("start"))
-	for i := start; i < start+5; i++ {
-		_, i2 := hs.ExampleScrape(page, i)
-		fmt.Printf("第%d页", i2)
-	}
-	c.JSON(http.StatusOK, gin.H{
-		"page":  page,
-		"start": start,
-		"msg":   "操作成功",
-		"url":   "http://localhost:8500/getData/" + strconv.Itoa(page) + "/" + strconv.Itoa(start+5),
-	})
 }
